ml_research/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ml-research"
version = "0.1.0"
description = "Machine learning research for churn prediction"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "jupyter>=1.0.0",
    "snowflake-connector-python[pandas]>=3.12.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "pyarrow>=17.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/churn_research"]

[tool.ruff]
target-version = "py311"
line-length = 119
show-fixes = true
src = ["src"]
extend-exclude = [
    "local/",
    "docs/",
    "build/",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "C4",  # flake8-comprehensions
    "PIE", # flake8-pie
    "TCH", # flake8-type-checking
    "PTH", # flake8-use-pathlib
    "ERA", # eradicate
    "PD",  # pandas-vet
    "PL",  # Pylint
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501",
    "PLR0913",
    "PLR0912",
    "PLR0915",
    "UP015",
    "UP045",
    "ERA001",
    "RUF001",
    "PLW2901",
    "PLC0415",
    "PLR2004",
    "PTH123",
]

[tool.ruff.lint.pep8-naming]
classmethod-decorators = [
    "classmethod",
    "pydantic.validator",
    "pydantic.field_validator",
]

[tool.ruff.lint.isort]
force-sort-within-sections = true
split-on-trailing-comma = true
combine-as-imports = true
force-wrap-aliases = true
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
